Update beagleboard.org-4gb.sh
This is mostly a hack to see if it will trigger a build. I intend to make a systemd service to listen to socket 1880 and fork node-red to add several Bone specific nodes.

It seems awkward to have so many scripts look so similar, but it is fine if these are really the things that will be customized per image.